+++ mercury-service/src/main/java/org/myapp/mercury/app/service/LogisticService.java
@@ -18,7 +18,7 @@ public interface LogisticService {
 
 	List<Supplier> findSuppliers();
 
-	Optional<Supplier> findSupplierById(int id);
+	Optional<Supplier> findSupplierById(long id);
 
 	Optional<List<Supplier>> findSupplierByName(String name);
 
@@ -26,13 +26,13 @@ public interface LogisticService {
 
 	List<Supply> findSuppliesByCriteria(SupplyCriteria criteria, RangeCriteria rangeCriteria);
 
-	void deleteSupplier(int id);
+	void deleteSupplier(long id);
 
 	void updateSupplier(Supplier supplier);
 
-	void saveSupply(int supplierId, Supply supply);
+	void saveSupply(long supplierId, Supply supply);
 
-	void deleteSupply(int id);
+	void deleteSupply(long id);
 
 	void updateSupply(Supply supply);
